Fixing folder structures in CMakeLists.txt and eliminating common.h in all files

+++ Src/DasherCore/ButtonMultiPress.cpp
@@ -28,7 +28,7 @@ CButtonMultiPress::CButtonMultiPress(CSettingsStore* pSettingsStore, CDasherInte
 }
 
 void CButtonMultiPress::Timer(unsigned long iTime, CDasherView *pView, CDasherInput *pInput, CDasherModel *pModel, CExpansionPolicy **pol) {
-  if(m_bKeyDown && !m_bKeyHandled && ((iTime - m_iKeyDownTime) > m_pSettingsStore->GetLongParameter(LP_HOLD_TIME))) {
+  if(m_bKeyDown && !m_bKeyHandled && ((iTime - m_iKeyDownTime) > static_cast<unsigned long>(m_pSettingsStore->GetLongParameter(LP_HOLD_TIME)))) {
     ButtonEvent(iTime, m_iHeldId, 1, pModel);
     m_bKeyHandled = true;
   }
@@ -41,7 +41,7 @@ void CButtonMultiPress::KeyDown(unsigned long iTime, Keys::VirtualKey Key, CDash
 
   // Check for multiple clicks
   if(Key == m_iQueueId && !m_deQueueTimes.empty()) {
-    if ( (iTime - m_deQueueTimes.back()) > m_pSettingsStore->GetLongParameter(LP_MULTIPRESS_TIME) )
+    if ( (iTime - m_deQueueTimes.back()) > static_cast<unsigned long>(m_pSettingsStore->GetLongParameter(LP_MULTIPRESS_TIME)))
       m_deQueueTimes.clear(); //and fall through to record+process normally, below
     else
     {
